% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_marginal_rates.R
\name{compute_marginal_rates}
\alias{compute_marginal_rates}
\title{Computes Marginal Probabilities}
\usage{
compute_marginal_rates(fit, tibble = TRUE)
}
\arguments{
\item{fit}{The regression model computed by \code{glm}
with \code{binomial()} family.}

\item{tibble}{Logical. Should the resulting marginal means
and comparisons be returned as tibbles (\code{TRUE}, default)
or as raw {emmeans} tables (\code{FALSE})?}
}
\value{
A list
}
\description{
Given a logistic regression model, extracts all marginal
mean probabilities of the outcome as well as all pairwise
comparisons and associated p-values without adjustment
for multiple comparisons.
}
