% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Visualises a Heat Map}
\usage{
draw_heatmap(tab, data, rts, help, pal = "warm")
}
\arguments{
\item{tab}{A table of regression coefficients and accompanying
information as generated by \code{summarise_regressions}.}

\item{data}{A data.frame with raw data generated by \code{import_data}.}

\item{rts}{A character vector of response time variables to be log-transformed
as generated by \code{extract_rt_variables}.}

\item{help}{A list with helper files generated by \code{extract_helpers}.}

\item{pal}{A character indicating which colour scale to use. Either
red tiles with blue asterisks (\code{"warm"} or \code{"sun"}, default) or
blue tiles with red asterisks (\code{"cold"} or \code{"sea"}).}
}
\value{
A list with two components:
\describe{
\item{\code{heatmap}}{A {ggplot2} object containing a heatmap
of all hippocampus structure/cognitive index available based on
s values of the hippocampus * diagnosis interaction from a multiple
regression analysis of the cognitive profile.}
\item{\code{scatters}}{A gtable object containing scatter dot plot with
cognition on ordinate and standardised hippocampi structure volume estimates
on abscissa. Includes only cognition/hippocampus pairs with s > 5 fot the
interaction term in a multiple regression of cognition on dignosis, hippocampus,
and additive covariates.}
}
}
\description{
Uses results of the regression of cognition on hippocampal
structures. The object of this visualisation are s values with
a threshold s > 5 to signify potentially interesting relations.
}
\details{
Whereas the \code{heatmap} will render a plot simply on calling it,
the scatter dot plots need to be plotted explicitly via
\code{plot(fig$scatters)} (see examples).
}
\examples{
\dontrun{
# Paths to files:
paths <- unlist(c(data_paths(), helper_paths())) |> purrr::keep(\(x) grepl("\\\\.", x))

# Helpers:
H <- paths |>
  helper_paths() |>
  extract_helpers()

# Data:
D <- paths |>
  data_paths() |>
  import_data(H)

# Response time variables:
RTs <- extract_rt_variables(H)

# Response summaries:
sums <- fit_regressions(
    preprocess_data(D, H, RTs, "df"),
    setup_regressions(H),
    H
  ) |> summarise_regressions()

tab <- sums$`cognition|hippocampi`

# Run the function:
figs <- draw_heatmap(tab, D, RTs, H, "warm")
figs1 <- draw_heatmap(tab, D, RTs, H, "sun") # the same as above
figs2 <- draw_heatmap(tab, D, RTs, H, "sea") # flipped colours

# Show the plots:
figs$heatmap
plot(figs$scatter) # need to plot()
}

}
\seealso{
\code{\link[=summarise_regressions]{summarise_regressions()}}, \code{\link[=extract_helpers]{extract_helpers()}}, \code{\link[=import_data]{import_data()}},
\code{\link[=extract_rt_variables]{extract_rt_variables()}}.
}
