% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{extract_helpers}
\alias{extract_helpers}
\title{Prepare 'Helper' Files}
\usage{
extract_helpers(paths)
}
\arguments{
\item{paths}{Paths to files.}
}
\value{
A list with data frames containing different
kinds of helper files.
}
\description{
This function needs to be adjusted manually
if desired.
It extract helper files that are later used to
control analysis flow.
}
