% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{extract_marginal_probabilities}
\alias{extract_marginal_probabilities}
\title{Extract Marginal Probabilities}
\usage{
extract_marginal_probabilities(fit)
}
\arguments{
\item{fit}{The regression model computed by \code{glm}.}
}
\value{
A list.
}
\description{
From a logistic regression, extracts all marginal
probabilities and all pairwise comparisons via
the \pkg{emmeans}.
}
