% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_funs.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit Logistic Regressions}
\usage{
fit_glm(d, outcomes, X = "SUBJ * AHI.F * GENDER", w = FALSE)
}
\arguments{
\item{d}{The data.}

\item{outcomes}{A character vector of the outcomes
of interest.}

\item{X}{The right side of the linear model.}

\item{w}{Indicator whether regression weights shall
be used (TRUE) or not(FALSE, default)}
}
\value{
A list with logistic regression models,
one per outcome
}
\description{
Loops through outcomes to calculate a set of
logistic regressions.
}
\seealso{
\code{\link[=fit_regressions]{fit_regressions()}} wraps this function.
}
