% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{helper_paths}
\alias{helper_paths}
\title{Document Paths to Helper Files}
\usage{
helper_paths(tracker)
}
\arguments{
\item{tracker}{A character vector. In the \pkg{targets}
pipeline serves to control the flow if the helper
files change.}
}
\value{
A data.frame containing paths to helpers
}
\description{
This function needs to be adjusted manually if
desired. It is used in the \pkg{targets} pipeline
to show helper files locations when importing them.
}
