% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_regressions.R
\name{setup_regressions}
\alias{setup_regressions}
\title{Set-up Regressions}
\usage{
setup_regressions(help)
}
\arguments{
\item{help}{A list with helpers files prepared
by \code{extract_helpers}.}
}
\value{
A list with linear models, one per outcome.
}
\description{
Prepares a set-up file for fitting regressions for
this project
}
\seealso{
\code{\link[=fit_regressions]{fit_regressions()}} makes use of the outcome
of this function.
}
